requirements.txt
streamlit
numpy 
css

opencv-python
easyocr 
pillow 
opencv-python
streamlit
